services/monitoring_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
psutil==5.9.6
prometheus-client==0.19.0
python-dotenv==1.0.0
boto3==1.29.3
aiohttp==3.9.1
python-multipart==0.0.6 